requirements.txt
streamlit
google-generativeai
PyPDF2
requests
beautifulsoup4